Initial support for time based OTP validation (#44)
* feat: initial support for time based OTP validation

The current implementation is naive in the sense that relies on the defaults of pyotp.
It has been verified to work with the "Google Authenticator".

* fix: removed unnecessary OTP token generation

* chore: initial support for OTP QR code generation

* chore: more generic 2FA auth approach

Allows more 2FA methods to be plugged in, and allows for more flexibility.
Should eventually allow FIDO2 authentication.

* chore: app name support in OTP QR Code

* feat: support for social login auth using 2FA

Redirect the user to a 2FA page after the login is successful.

* chore: small URL naming change

* fix: re-format issues

* chore: prevents user from run login via API if 2FA is enabled

Provides a proper security mechanism to prevent users from logging in via API if 2FA is enabled. This is a security measure to prevent unauthorized access to the account.